Moving types definitions for removing circular dependencies
