+++ packages/core/package.json
@@ -42,7 +42,6 @@
     "@soroban-react/freighter": "^9.0.2",
     "@soroban-react/types": "^9.0.2",
     "@soroban-react/xbull": "^1.0.1",
-    "@soroban-react/contracts": "^9.0.2",
     "@stellar/stellar-sdk": "11.1.0"
   },
   "devDependencies": {
