+++ packages/types/src/index.tsx
@@ -63,25 +63,3 @@ export type ContractDeploymentInfo = {
   networkPassphrase: string,
   contractAddress: string
 }
-
-import * as StellarSdk from '@stellar/stellar-sdk'
-
-export type WrappedContractInvokeArgs = {
-  // NO NEED contractAddress: string
-  method: string
-  args?: StellarSdk.xdr.ScVal[] | undefined
-  signAndSend?: boolean
-  fee?: number
-  skipAddingFootprint?: boolean
-  secretKey?: string
-  // NO NEED MAYBE sorobanContext: SorobanContextType
-  // If useSorobanReact called inside of wrapped function
-  reconnectAfterTx?: boolean
-}
-
-import {TxResponse} from '@soroban-react/contracts'
-
-export type WrappedContract = {
-  deploymentInfo: ContractDeploymentInfo,
-  invoke: (args: WrappedContractInvokeArgs) => Promise<TxResponse | StellarSdk.xdr.ScVal>
-}
